Refactor(config): Split the configuration validator
validator.py held 538 lines, past the 400-line reviewability limit the
aislop code-quality gate applies.

Split it into a package at the same dotted path: schema_errors (the
jsonschema message formatters), results (the validation result model),
core (ConfigValidator and validate_config_file) and reporting (the
stderr rendering). The package __init__ re-exports the previous public
names, so all import sites are unchanged.

The schema path in ConfigValidator.__init__ gains one parent level, as
the module now sits one directory deeper than config/schema.json.
